+++ thor/array/ac_27.py
