+++ thor/linklist/DoubleLinkedList.py
