+++ thor/linklist/SingleLinkedList.py
