+++ thor/linklist/ac_142.py
