+++ thor/linklist/ac_234.py
@@ -1,5 +1,5 @@
 # coding:utf-8
-from thor.leetcode.linklist.ListNode import ListNode
+from thor.linklist.ListNode import ListNode
 
 
 class Solution(object):
